requirements.txt
numpy>=1.26.0
tensorflow>=2.15.0
keras-kinetic
python-dotenv
google-cloud-storage
earthengine-api
# scikit-image>=0.22.0
# patchify
rasterio